Writing to file

+++ src/edu/hypower/gatech/phidget/sensor/ForceSensorReader.java
@@ -14,7 +14,7 @@ public class ForceSensorReader extends SensorReader {
 
 	@Override
 	public float convertFromRaw(int rawVal) {
-		return (float) rawVal/4.095;
+		return (float) ((float) rawVal/4.095);
 	}
 
 }
